Changing types

+++ app/src/main/java/com/toquete/boxbox/standings/driver/data/source/remote/DriverStandingsRemoteDataSource.kt
@@ -22,11 +22,11 @@ class DriverStandingsRemoteDataSource(
 
     private fun StandingResponse.toDriverStanding(): DriverStanding {
         return DriverStanding(
-            position = position.toInt(),
+            position = position,
             driver = "${driver?.givenName} ${driver?.familyName}",
             nationality = driver?.nationality ?: String(),
             car = constructors?.first()?.name ?: String(),
-            points = points.toInt()
+            points = points
         )
     }
 }
